+++ platform-api/src/main/java/com/platform/entity/ProductVo.java
@@ -16,9 +16,9 @@ public class ProductVo extends BaseEntity  implements Serializable {
     private static final long serialVersionUID = 1L;
 
     /**商品Id**/
-    private Integer goods_id;
+    private Long goods_id;
     /**产品Id**/
-    private Integer product_id;
+    private Long product_id;
     /**商品规格ids**/
     private String goods_specification_ids;
     /**商品序列号**/
